Fix diff types comparation warning

+++ utils.cpp
@@ -1,3 +1,4 @@
+#include <cstddef>
 #include <memory>
 #include <vector>
 using namespace std;
@@ -56,12 +57,12 @@ vector<shared_ptr<Node<S, A>>> expands(Problem<S, A> &problem, shared_ptr<Node<S
 }
 
 // Check if position is at bounds of the table
-bool at_bounds(int x, int y, size_t table_size) {
+bool at_bounds(size_t x, size_t y, size_t table_size) {
   return x >= 0 && x < table_size &&
          y >= 0 && y < table_size;
 }
 
-bool at_bounds(pair<int, int> position, size_t table_size) {
+bool at_bounds(pair<size_t, size_t> position, size_t table_size) {
   return position.first >= 0 && position.first < table_size &&
          position.second >= 0 && position.second < table_size;
 }
